email des comptes avec un `concerne` plutôt qu'un `observer`
Cet commit adresse l'erreur rollbar numéro 224 : les emails de création
de compte n'était pas toujours envoyé, dans le cas ou le mailer tentait
d'envoyer le mail alors que la transaction de création de compte ne
s'était pas encore terminée.

Les `observers` ne permettent pas d'attacher des callback après la
transaction mais seulement pour la création.
Je suis passé à l'utilisation des `Concerns` qui est apparemment un
concept natif de `ActiveSupport` depuis Rails 4.
http://vaidehijoshi.github.io/blog/2015/10/13/stop-worrying-and-start-being-concerned-activesupport-concerns/
