+++ Gemfile
@@ -42,7 +42,6 @@ gem 'mini_racer', '~> 0.4.0'
 gem 'pg', '~> 1.0'
 gem 'rack-cors', require: 'rack/cors'
 gem 'rails-i18n'
-gem 'rails-observers'
 gem 'recipient_interceptor'
 gem 'redcarpet'
 gem 'redis'
