+++ config/application.rb
@@ -40,8 +40,6 @@ module EvaServeur
       g.orm :active_record, foreign_key_type: :uuid
     end
 
-    config.active_record.observers = :compte_observer
-
     config.action_controller.asset_host = "#{ENV['PROTOCOLE_SERVEUR']}://#{ENV['HOTE_SERVEUR']}"
     config.action_mailer.asset_host = "#{ENV['PROTOCOLE_SERVEUR']}://#{ENV['HOTE_SERVEUR']}"
     Rails.application.routes.default_url_options = {
